requirements.txt
# ============================================================
# Multi-Agent Stock Forecasting System — Requirements
# Python 3.10+
# ============================================================

# ── Core Scientific Stack ────────────────────────────────────
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# ── PyTorch (CPU build; swap for CUDA wheel from pytorch.org) ─
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# ── Google TimesFM ───────────────────────────────────────────
timesfm>=1.0.0

# ── Amazon Chronos ───────────────────────────────────────────
chronos-forecasting>=1.2.0

# ── HuggingFace ecosystem (shared by TimesFM + Chronos) ──────
transformers>=4.40.0
accelerate>=0.28.0
huggingface-hub>=0.21.0
tokenizers>=0.15.0

# ── Data Sources ─────────────────────────────────────────────
yfinance>=0.2.36
requests>=2.31.0

# ── Config ───────────────────────────────────────────────────
PyYAML>=6.0.1

# ── NLP / Sentiment ──────────────────────────────────────────
nltk>=3.8.1
vaderSentiment>=3.3.2

# ── Reliability ──────────────────────────────────────────────
tenacity>=8.2.3

# ── Progress & Utilities ────────────────────────────────────
tqdm>=4.66.0
python-dateutil>=2.8.2
pytz>=2024.1

# ── Visualisation ───────────────────────────────────────────
plotly>=5.18.0
# matplotlib>=3.7.0
